LayoutTests/media/modern-media-controls/controls-bar/controls-bar-stacking-context-expected.txt

Testing ControlsBar establishes a stacking context.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(controlsBar.element).willChange is "z-index"
PASS successfullyParsed is true

TEST COMPLETE
